+++ src/v0.8/module/dataset/library/DatasetMetadataLIB.sol
@@ -74,7 +74,7 @@ library DatasetMetadataLIB {
         string memory _description,
         string memory _source,
         string memory _accessMethod,
-        uint256 _sizeInBytes,
+        uint64 _sizeInBytes,
         bool _isPublic,
         uint64 _version
     ) external {
@@ -95,7 +95,7 @@ library DatasetMetadataLIB {
         self.metadata.source = _source;
         self.metadata.accessMethod = _accessMethod;
         self.metadata.submitter = msg.sender;
-        self.metadata.createdBlockNumber = block.number;
+        self.metadata.createdBlockNumber = uint64(block.number);
         self.metadata.sizeInBytes = _sizeInBytes;
         self.metadata.isPublic = _isPublic;
         self.metadata.version = _version;
@@ -119,8 +119,8 @@ library DatasetMetadataLIB {
             string memory source,
             string memory accessMethod,
             address submitter,
-            uint256 createdBlockNumber,
-            uint256 sizeInBytes,
+            uint64 createdBlockNumber,
+            uint64 sizeInBytes,
             bool isPublic,
             uint64 version
         )
